Renaming components with lowercase kebab style
