+++ src/components/EventBox/request-view.tsx
